requirements.txt
Flask
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
Flask-Login